backend/requirements.txt
# Core FastAPI dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
alembic>=1.13.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
aiosqlite>=0.19.0

# Pydantic
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Security
bcrypt>=4.1.2
python-jose[cryptography]>=3.3.0
passlib>=1.7.4

# HTTP Client
httpx>=0.26.0

# LLM - Hugging Face
transformers>=4.40.0
torch>=2.2.0
accelerate>=0.27.0
sentencepiece>=0.2.0
sentence-transformers>=2.3.0
bitsandbytes>=0.42.0

# File Processing
aiofiles>=23.2.1
python-docx>=1.1.0
pdfplumber>=0.10.3
python-magic>=0.4.27
python-magic-bin>=0.4.14; sys_platform == 'win32'

# Logging
structlog>=24.1.0

# LLM Service - llama.cpp (Phase 10 local testing)
llama-cpp-python>=0.2.0

# Multi-Agent Orchestration - LangGraph
langgraph>=0.0.40
langchain-core>=0.1.30

# vLLM (Production, optional for Phase 10)
# vllm>=0.3.0
